+++ GalaxyWorld.Core/Models/Star/StarInsert.cs
@@ -6,26 +6,26 @@ public class StarInsert
 {
     public int? Constellation { get; init; }
     public string? ProperName { get; init; }
-    public required decimal RightAscension { get; init; }
-    public required decimal Declination { get; init; }
+    public required double RightAscension { get; init; }
+    public required double Declination { get; init; }
     public required string PosSrc { get; init; }
-    public decimal? Distance { get; init; }
-    public decimal? X0 { get; init; }
-    public decimal? Y0 { get; init; }
-    public decimal? Z0 { get; init; }
+    public double? Distance { get; init; }
+    public double? X0 { get; init; }
+    public double? Y0 { get; init; }
+    public double? Z0 { get; init; }
     public string? DistanceSrc { get; init; }
-    public required decimal Magnitude { get; init; }
-    public decimal? AbsoluteMagnitude { get; init; }
-    public decimal? ColourIndex { get; init; }
+    public required double Magnitude { get; init; }
+    public double? AbsoluteMagnitude { get; init; }
+    public double? ColourIndex { get; init; }
     public required string MagnitudeSrc { get; init; }
-    public decimal? RadialVelocity { get; init; }
+    public double? RadialVelocity { get; init; }
     public string? RadialVelocitySrc { get; init; }
-    public decimal? ProperMotionRightAscension { get; init; }
-    public decimal? ProperMotionDeclination { get; init; }
+    public double? ProperMotionRightAscension { get; init; }
+    public double? ProperMotionDeclination { get; init; }
     public string? ProperMotionSrc { get; init; }
-    public decimal? VelocityX { get; init; }
-    public decimal? VelocityY { get; init; }
-    public decimal? VelocityZ { get; init; }
+    public double? VelocityX { get; init; }
+    public double? VelocityY { get; init; }
+    public double? VelocityZ { get; init; }
     public string? SpectralType { get; init; }
     public string? SpectralTypeSrc { get; init; }
     public IEnumerable<CatalogueEntryInsertWithStar>? CatalogueEntries { get; init; }
